+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - GP2-3074 - pipeline vfm survey add questions object
 
 ### Fixed bugs
+- GP2-3260 - Password reset validation fix
 - GP2-3017 - Added test for partial user profile update
 - GP2-2445 - JSON widget for VFM answer admin
 - GP2-2867 - Dockerise d-sso